+++ udata/tests/dataset/test_dataset_model.py
@@ -390,10 +390,18 @@ class LicenseModelTest:
         assert isinstance(found, License)
         assert license.id == found.id
 
+    def test_exact_match_by_url_in_string_real_world(self):
+        license = LicenseFactory(url="http://www.data.gouv.fr/Licence-Ouverte-Open-Licence")
+        found = License.guess(
+            "Licence Ouverte 1.0 http://www.data.gouv.fr/Licence-Ouverte-Open-Licence."
+        )
+        assert isinstance(found, License)
+        assert license.id == found.id
+
     def test_exact_match_by_url_in_string_two_urls(self):
         license = LicenseFactory()
         found = License.guess(
-            f"Here is my license: {license.url} and another link: https://example.com/license"
+            f"Here is my license: {license.url} and another link: https://example.com/example"
         )
         assert isinstance(found, License)
         assert license.id == found.id